third_party/WebKit/LayoutTests/fast/css/cssom-missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
var style = document.body.style;
shouldBeType("style", "CSSStyleDeclaration");
shouldThrow("style.item()");
shouldThrow("style.getPropertyValue()");
shouldThrow("style.getPropertyPriority()");
shouldThrow("style.setProperty()");
shouldThrow("style.setProperty('color')");
shouldThrow("style.removeProperty()");
</script>
</body>
</html>
