third_party/blink/web_tests/fast/css/cssom-missing-arguments-expected.txt

PASS style is an instance of CSSStyleDeclaration
PASS style.item() threw exception TypeError: Failed to execute 'item' on 'CSSStyleDeclaration': 1 argument required, but only 0 present..
PASS style.getPropertyValue() threw exception TypeError: Failed to execute 'getPropertyValue' on 'CSSStyleDeclaration': 1 argument required, but only 0 present..
PASS style.getPropertyPriority() threw exception TypeError: Failed to execute 'getPropertyPriority' on 'CSSStyleDeclaration': 1 argument required, but only 0 present..
PASS style.setProperty() threw exception TypeError: Failed to execute 'setProperty' on 'CSSStyleDeclaration': 2 arguments required, but only 0 present..
PASS style.setProperty('color') threw exception TypeError: Failed to execute 'setProperty' on 'CSSStyleDeclaration': 2 arguments required, but only 1 present..
PASS style.removeProperty() threw exception TypeError: Failed to execute 'removeProperty' on 'CSSStyleDeclaration': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
